Add .ipynb notebook
The notebook includes R code to create 4 functions to (i) select the most dissimilar combinations for each sequence, (ii) view the distance matrixes, (iii) get an ordered dataframe with the best combinations and (iv) plot the nucleotide frequencies of a given combination. It also includes two examples on how to use the functions.